Rename updateGlobalState to setGlobalState - Also includes some re-organizing of core files
