+++ packages/core/index.js
@@ -1,20 +1,19 @@
-export * from "./src/js/Collection";
-export * from "./src/js/CollectionEntry";
-export * from "./src/js/Breakpoint";
-export * from "./src/js/FocusTrap";
-
 export * from "./src/js/utilities/toCamel";
 export * from "./src/js/utilities/toKebab";
 export * from "./src/js/utilities/toMilliseconds";
 
 export * from "./src/js/helpers/cssVar";
+export * from "./src/js/helpers/getConfig";
+export * from "./src/js/helpers/getCustomProps";
+export * from "./src/js/helpers/getElement";
 export * from "./src/js/helpers/getPrefix";
+export * from "./src/js/helpers/transition";
+export * from "./src/js/helpers/setGlobalState";
 
-export * from "./src/js/getConfig";
-export * from "./src/js/getCustomProps";
-export * from "./src/js/getElement";
+export * from "./src/js/Breakpoint";
+export * from "./src/js/Collection";
+export * from "./src/js/CollectionEntry";
+export * from "./src/js/FocusTrap";
 export * from "./src/js/localStore";
 export * from "./src/js/teleport";
 export * from "./src/js/themeStore";
-export * from "./src/js/transition";
-export * from "./src/js/updateGlobalState";
