+++ packages/core/src/js/helpers/getConfig.js
