+++ packages/core/src/js/helpers/transition.js
@@ -1,5 +1,5 @@
 // import { cssVar } from "./cssVar";
-import { toMilliseconds } from "./utilities/toMilliseconds";
+import { toMilliseconds } from "../utilities/toMilliseconds";
 
 /**
  * Get the value of a CSS custom property (variable).
