+++ packages/core/tests/updateGlobalState.test.js
@@ -1,4 +1,4 @@
-import { updateGlobalState } from "../index";
+import { setGlobalState } from "../index";
 import "@testing-library/jest-dom/vitest";
 
 document.body.innerHTML = `
@@ -11,14 +11,14 @@ document.body.innerHTML = `
 const header = document.querySelector(".header");
 const main = document.querySelector(".main");
 
-describe("updateGlobalState()", () => {
+describe("setGlobalState()", () => {
   it("should apply inert and aria hidden to passed selectors", () => {
     expect(main.inert).not.toBe(true);
     expect(header.inert).not.toBe(true);
     expect(main).not.toHaveAttribute("aria-hidden");
     expect(header).not.toHaveAttribute("aria-hidden");
 
-    updateGlobalState(true, ".main, .header", "body, .main");
+    setGlobalState(true, ".main, .header", "body, .main");
     expect(main.inert).toBe(true);
     expect(header.inert).toBe(true);
     expect(main).toHaveAttribute("aria-hidden", "true");
@@ -26,7 +26,7 @@ describe("updateGlobalState()", () => {
   });
 
   it("should remove inert and aria hidden when set to false", () => {
-    updateGlobalState(false, ".main, .header", "body, .main");
+    setGlobalState(false, ".main, .header", "body, .main");
     expect(main.inert).not.toBe(true);
     expect(header.inert).not.toBe(true);
     expect(main).not.toHaveAttribute("aria-hidden");
@@ -34,7 +34,7 @@ describe("updateGlobalState()", () => {
   });
 
   it("should do nothing if selector is not passed", () => {
-    updateGlobalState(true, null, null);
+    setGlobalState(true, null, null);
     expect(main.inert).not.toBe(true);
     expect(header.inert).not.toBe(true);
     expect(main).not.toHaveAttribute("aria-hidden");
@@ -44,19 +44,19 @@ describe("updateGlobalState()", () => {
   it("should apply overflow hidden to passed selectors", () => {
     expect(document.body).not.toHaveStyle("overflow: hidden");
     expect(main).not.toHaveStyle("overflow: hidden");
-    updateGlobalState(true, ".main, .header", "body, .main");
+    setGlobalState(true, ".main, .header", "body, .main");
     expect(document.body).toHaveStyle("overflow: hidden");
     expect(main).toHaveStyle("overflow: hidden");
   });
 
   it("should remove overflow hidden when set to false", () => {
-    updateGlobalState(false, ".main, .header", "body, .main");
+    setGlobalState(false, ".main, .header", "body, .main");
     expect(document.body).not.toHaveStyle("overflow: hidden");
     expect(main).not.toHaveStyle("overflow: hidden");
   });
 
   it("should do nothing if selector is not passed", () => {
-    updateGlobalState(true, null, null);
+    setGlobalState(true, null, null);
     expect(document.body).not.toHaveStyle("overflow: hidden");
     expect(main).not.toHaveStyle("overflow: hidden");
   });
