+++ packages/drawer/src/js/close.js
@@ -1,4 +1,4 @@
-import { transition, updateGlobalState } from "@vrembem/core";
+import { transition, setGlobalState } from "@vrembem/core";
 import { updateFocusState, getDrawer } from "./helpers";
 
 export async function close(query, transitionOverride, focus = true) {
@@ -31,7 +31,7 @@ export async function close(query, transitionOverride, focus = true) {
     entry.state = "closed";
 
     // Update the global state if mode is modal.
-    if (entry.mode === "modal") updateGlobalState(false, entry.getSetting("selectorInert"), entry.getSetting("selectorOverflow"));
+    if (entry.mode === "modal") setGlobalState(false, entry.getSetting("selectorInert"), entry.getSetting("selectorOverflow"));
 
     // Set focus to the trigger element if the focus param is true.
     if (focus) {
